id_verify_api/app/services.py
import uuid
from typing import List, Dict
import base64
from io import BytesIO
from PIL import Image
from .models import (
    VerificationRequest,
    VerificationResult,
    VerificationStatus,
    DocumentSubmission
)

class VerificationService:
    def __init__(self):
        self.verifications: Dict[str, VerificationResult] = {}

    def validate_document_image(self, image_base64: str) -> bool:
        """Validate if the provided image is valid and meets quality standards."""
        try:
            # Decode base64 image
            image_data = base64.b64decode(image_base64)
            img = Image.open(BytesIO(image_data))
            
            # Basic image validation
            min_width, min_height = 1000, 1000
            if img.width < min_width or img.height < min_height:
                return False
                
            return True
        except Exception:
            return False

    def verify_document(self, document: DocumentSubmission) -> dict:
        """Verify a single document submission."""
        # Validate document images
        if not self.validate_document_image(document.document_front):
            return {
                "is_valid": False,
                "errors": ["Document front image is invalid or low quality"]
            }

        if document.document_back and not self.validate_document_image(document.document_back):
            return {
                "is_valid": False,
                "errors": ["Document back image is invalid or low quality"]
            }

        # In a real implementation, you would:
        # 1. Call OCR service to extract text from documents
        # 2. Validate extracted information against provided data
        # 3. Check document authenticity features
        # 4. Verify document with issuing authority's database

        # Simulated verification result
        return {
            "is_valid": True,
            "confidence_score": 0.95,
            "verified_fields": ["document_number", "expiry_date"]
        }

    def verify_identity(self, request: VerificationRequest) -> VerificationResult:
        """Process a complete identity verification request."""
        verification_id = str(uuid.uuid4())
        
        # Verify each submitted document
        verified_fields = []
        failed_fields = []
        total_confidence = 0.0
        
        for doc in request.documents:
            result = self.verify_document(doc)
            if result["is_valid"]:
                verified_fields.extend(result["verified_fields"])
                total_confidence += result["confidence_score"]
            else:
                failed_fields.extend(result["errors"])

        # Calculate overall confidence score
        confidence_score = total_confidence / len(request.documents) if request.documents else 0

        # Determine verification status
        status = (
            VerificationStatus.APPROVED if confidence_score > 0.8
            else VerificationStatus.NEEDS_REVIEW if confidence_score > 0.5
            else VerificationStatus.REJECTED
        )

        result = VerificationResult(
            verification_id=verification_id,
            status=status,
            confidence_score=confidence_score,
            verified_fields=verified_fields,
            failed_fields=failed_fields
        )

        self.verifications[verification_id] = result
        return result

    def get_verification_status(self, verification_id: str) -> VerificationResult:
        """Retrieve the status of a verification request."""
        return self.verifications.get(verification_id)